+++ packages/@cldcvr/flow-lineage/src/components/f-lineage/draw/draw-lineage.ts
@@ -3,6 +3,6 @@ import drawLinks from "./draw-links";
 import drawNodes from "./draw-nodes";
 
 export default function drawLineage(params: DrawLineageParams) {
-  drawNodes(params);
   drawLinks(params);
+  drawNodes(params);
 }